requirements.txt
dataclasses_json==0.5.7
discord.py==2.1.0
python-dotenv==0.21.0
requests==2.28.1
